landing page: hero component and highlight section
